Changes for the hyrax-1.10 release.
I've made a few changes for the Hyrax 1.10 release. First, I dropped
the gdal version required to 1.7.3 so that I can build rpms that will
work with the package as it appears on CentOS 6. Two of the four tests
fail and I'm going to mark those as expect fail in the next commit.